refactor: logger.logger
Make logger.logger as a custom logger to keep its functionality after
upgrading to `go-logr/logr` v1.x.
